Change rest to grpc for performance

+++ src/features/auth/service/refresh-token.service.ts
@@ -9,6 +9,8 @@ import { AppLogger } from '@shared/logger';
 import { basename } from 'path';
 import { GetUserTokenQueryDto, LogOutRequestDto } from '@auth/dto/request.dto';
 import { RotateTokenResponseDto } from '@auth/dto/response.dto';
+import { status } from '@grpc/grpc-js';
+
 @Injectable()
 export class RefreshTokenService {
   constructor(
@@ -16,11 +18,11 @@ export class RefreshTokenService {
     private jwtService: JwtService,
     private appLogger: AppLogger,
   ) {}
+
   async issueTokenPair({ userId }) {
     const refreshToken = v4();
     const tokenHash = await bcrypt.hash(refreshToken, 10);
     const expiresAt = new Date(Date.now() + 7 * 24 * 60 * 60 * 1000);
-
     const accessToken = this.jwtService.sign({ userId }, { expiresIn: '5m' });
     return {
       accessToken,
@@ -42,30 +44,34 @@ export class RefreshTokenService {
       .addMsgParam('rotateToken');
     this.appLogger.log(`WILL rotate token for user ${userId}`);
 
-    const record = await this.refreshTokenRepository.findOne({
-      userId,
-      sessionId,
-    });
+    const record = await this.refreshTokenRepository.findOne({ userId, sessionId });
     if (!record) {
-      throw new RefreshTokenError('No token found');
+      throw new RefreshTokenError({
+        errorCode: status.NOT_FOUND,
+        details: 'No token found',
+      });
     }
 
     const validToken = await bcrypt.compare(refreshToken, record.tokenHash);
     if (!validToken) {
-      throw new RefreshTokenError('Invalid token');
+      throw new RefreshTokenError({
+        errorCode: status.UNAUTHENTICATED,
+        details: 'Invalid token',
+      });
     }
 
     const checkUsedTokenResult = await Promise.all(
-      record.usedTokenHashes.map((token) =>
-        bcrypt.compare(refreshToken, token),
-      ),
+      record.usedTokenHashes.map((token) => bcrypt.compare(refreshToken, token)),
     );
     if (checkUsedTokenResult.some((item) => item)) {
       await this.refreshTokenRepository.deleteMany({ userId });
       this.appLogger.error(
         `FAILED rotate token for user ${userId}: rotate an used token detected`,
       );
-      throw new RefreshTokenError('Invalid credential, please login again');
+      throw new RefreshTokenError({
+        errorCode: status.UNAUTHENTICATED,
+        details: 'Invalid credential, please login again',
+      });
     }
 
     const isTokenExpired = record.expiresAt < new Date();
@@ -74,17 +80,16 @@ export class RefreshTokenService {
       this.appLogger.error(
         `FAILED rotate token for user ${userId}: rotate an expired token`,
       );
-      throw new RefreshTokenError('Refresh token expired, please login again');
+      throw new RefreshTokenError({
+        errorCode: status.UNAUTHENTICATED,
+        details: 'Refresh token expired, please login again',
+      });
     }
 
-    const { accessToken, refreshTokenInfo } = await this.issueTokenPair({
-      userId,
-    });
-
+    const { accessToken, refreshTokenInfo } = await this.issueTokenPair({ userId });
     record.usedTokenHashes.push(record.tokenHash);
     record.expiresAt = refreshTokenInfo.expiresAt;
     record.tokenHash = refreshTokenInfo.tokenHash;
-
     await record.save();
     this.appLogger.log(`DID rotate token for user ${userId}: successfully`);
     return {
@@ -99,11 +104,9 @@ export class RefreshTokenService {
       .addLogContext(context.traceId)
       .addMsgParam(basename(__filename))
       .addMsgParam('revokeToken');
-
     this.appLogger.log(`WILL revoke token for user ${userId}`);
     await this.refreshTokenRepository.deleteMany({ userId });
     this.appLogger.log(`DID revoke token for user ${userId}`);
-
     return true;
   }
 
@@ -115,7 +118,9 @@ export class RefreshTokenService {
     this.appLogger.log(`WILL get token for user ${userId}`);
     const result = await this.refreshTokenRepository.findMany({ userId });
     this.appLogger.log(`DID get token for user ${userId}`);
-    return result;
+    return {
+      tokens: result.map((data) => data.userId),
+    };
   }
 
   async logOut(context: AppContext, { userId, sessionId }: LogOutRequestDto) {
@@ -126,10 +131,7 @@ export class RefreshTokenService {
     this.appLogger.log(
       `WILL log out for user ${userId} - session ${sessionId}`,
     );
-    const result = await this.refreshTokenRepository.deleteOne({
-      userId,
-      sessionId,
-    });
+    const result = await this.refreshTokenRepository.deleteOne({ userId, sessionId });
     this.appLogger.log(
       `DID log out for user ${userId} - session ${sessionId}}`,
     );
